requirements.txt
streamlit
langchain-google-genai
langchain-community
langchain-core
python-docx
python-dotenv
faiss-cpu
sentence-transformers
huggingface-hub
PyPDF2